#feat 02/21 1) 제품(Item) 엔티티와 연결된 ItemType, GlassSpec 을 관리자가 수정가능한 REST API 구현 (클라이언트 응답 정상 확인)
내일 할일 : 전체적인 테스트 코드 재검점
(DTO, 및 Entity 일부 수정으로 테스트 코드 파라미터 값에 컴파일 오류 발생한 부분 재점검)

+++ src/main/java/com/hyeobjin/web/admin/item/type/AdminItemTypeController.java
@@ -35,17 +35,22 @@ public class AdminItemTypeController {
         return ResponseEntity.status(HttpStatus.CREATED).build();
     }
 
-    @PostMapping("/update/{itemTypeId}")
+    @PostMapping("/update")
     @Operation(summary = "관리자 제품 타입 수정 API", description = "관리자가 기존 제품의 타입을 수정하기 위한 API 입니다.")
-    public ResponseEntity<Void> update(@PathVariable("itemTypeId") Long itemTypeId, @RequestBody UpdateItemTypeDTO updateItemTypeDTO) {
-        adminItemTypeService.updateItemType(itemTypeId, updateItemTypeDTO);
-        return ResponseEntity.noContent().build();
+    public ResponseEntity<String> update(@RequestBody UpdateItemTypeDTO updateItemTypeDTO) {
+        adminItemTypeService.updateItemType(updateItemTypeDTO);
+        return ResponseEntity.ok("제품 타입이 수정되었습니다.");
     }
 
     @DeleteMapping
     @Operation(summary = "관리자 제품 타입 삭제 API", description = "관리자가 기존 제품의 타입을 삭제하기 위한 API 입니다.")
-    public ResponseEntity<Void> delete(@RequestParam("itemTypeId") Long itemTypeId) {
+    public ResponseEntity<String> delete(@RequestParam("itemTypeIds") List<Long> itemTypeId) {
+
+        if (itemTypeId == null || itemTypeId.isEmpty()) {
+            return ResponseEntity.badRequest().body("삭제할 항목이 없습니다");
+        }
+
         adminItemTypeService.deleteItemType(itemTypeId);
-        return ResponseEntity.noContent().build();
+        return ResponseEntity.ok("성공적으로 삭제 되었습니다.");
     }
 }
